+++ game_patch/misc/player_fpgun.cpp
@@ -4,6 +4,7 @@
 #include <patch_common/AsmWriter.h>
 #include <common/config/BuildConfig.h>
 #include "alpine_options.h"
+#include "alpine_settings.h"
 #include "../rf/player/player.h"
 #include "../rf/player/camera.h"
 #include "../rf/sound/sound.h"
@@ -152,7 +153,7 @@ CodeInjection after_game_render_to_dynamic_textures{
 CallHook<void(rf::Matrix3&, rf::Vector3&, float, bool, bool)> player_fpgun_render_gr_setup_3d_hook{
     0x004AB411,
     [](rf::Matrix3& viewer_orient, rf::Vector3& viewer_pos, float horizontal_fov, bool zbuffer_flag, bool z_scale) {
-        horizontal_fov *= g_game_config.fpgun_fov_scale;
+        horizontal_fov *= g_alpine_game_config.fpgun_fov_scale;
         horizontal_fov = gr_scale_fov_hor_plus(horizontal_fov);
         player_fpgun_render_gr_setup_3d_hook
             .call_target(viewer_orient, viewer_pos, horizontal_fov, zbuffer_flag, z_scale);
@@ -163,10 +164,9 @@ ConsoleCommand2 fpgun_fov_scale_cmd{
     "r_fpgunfov",
     [](std::optional<float> scale_opt) {
         if (scale_opt) {
-            g_game_config.fpgun_fov_scale = std::clamp(scale_opt.value(), 0.1f, 1.5f);
-            g_game_config.save();
+            g_alpine_game_config.set_fpgun_fov_scale(scale_opt.value());
         }
-        rf::console::print("Fpgun FOV scale: {:.4f}", g_game_config.fpgun_fov_scale.value());
+        rf::console::print("Fpgun FOV scale: {:.4f}", g_alpine_game_config.fpgun_fov_scale);
     },
     "Set scale value applied to FOV setting for first person weapon models.",
     "r_fpgunfov [scale]",